requirements.txt
discord
openai
fastapi
black
isort
websockets
pydantic
langchain
uvicorn
jinja2
faiss-cpu
bs4
unstructured
libmagic
python-dotenv
